+++ src/app/layout.tsx
@@ -1,6 +1,7 @@
 import { Metadata } from 'next';
 import { Geist, Geist_Mono } from 'next/font/google';
 
+import TanStackQueryClientProvider from '@/components/query-client-provider';
 import { ThemeProvider } from '@/components/theme-provider';
 import { DEFAULT_METADATA } from '@/lib/metadata';
 import { Toaster } from '@/components/ui/sonner';
@@ -27,13 +28,15 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <html lang="en" suppressHydrationWarning>
-      <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>
-        <ThemeProvider attribute="class" defaultTheme="system" enableSystem disableTransitionOnChange>
-          {children}
-          <Toaster position="top-center" />
-        </ThemeProvider>
-      </body>
-    </html>
+    <TanStackQueryClientProvider>
+      <html lang="en" suppressHydrationWarning>
+        <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>
+          <ThemeProvider attribute="class" defaultTheme="system" enableSystem disableTransitionOnChange>
+            {children}
+            <Toaster position="top-center" />
+          </ThemeProvider>
+        </body>
+      </html>
+    </TanStackQueryClientProvider>
   );
 }